修正了紧张/压力/害怕的bug

+++ src/PokemonBattle.EffectsC/Moves/Attack/SpBasepower/HeavySlam.cs
@@ -7,7 +7,7 @@ using LightStudio.PokemonBattle.Data;
 
 namespace LightStudio.PokemonBattle.Game.Host.Effects.Moves
 {
-  class HeavySlam:AttackMoveE
+  class HeavySlam : AttackMoveE
   {
     public HeavySlam(int MoveId)
       : base(MoveId)
@@ -15,17 +15,12 @@ namespace LightStudio.PokemonBattle.Game.Host.Effects.Moves
 
     protected override void CalculateBasePower(DefContext def)
     {
-      int w = (int)(def.AtkContext.Attacker.Weight / def.Defender.Weight);
-      if (w >= 5)
-        def.BasePower = 120;
-      else if (w >= 4)
-        def.BasePower = 100;
-      else if (w >= 3)
-        def.BasePower = 80;
-      else if (w >= 2)
-        def.BasePower = 60;
-      else
-        def.BasePower = 40;
+      double w = def.AtkContext.Attacker.Weight / def.Defender.Weight;
+      if (w >= 5) def.BasePower = 120;
+      else if (w >= 4) def.BasePower = 100;
+      else if (w >= 3) def.BasePower = 80;
+      else if (w >= 2) def.BasePower = 60;
+      else def.BasePower = 40;
     }
   }
 }
